requirements.txt
dbt-postgres
streamlit
pandas
numpy
altair
sqlalchemy
psycopg2-binary
googleads
agate==1.9.1
altair==5.5.0
annotated-types==0.7.0
attrs==25.4.0
babel==2.17.0
blinker==1.9.0
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
daff==1.4.2
dbt-adapters==1.19.0
dbt-common==1.36.0
dbt-core==1.10.15
dbt-extractor==0.6.0
dbt-postgres==1.9.1
dbt-protos==1.0.382
dbt-semantic-interfaces==0.9.0
deepdiff==8.6.1
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.41.1
google-auth-oauthlib==1.2.3
googleads==48.0.0
greenlet==3.2.4
idna==3.11
importlib_metadata==8.7.0
isodate==0.6.1
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
leather==0.4.0
lxml==6.0.2
MarkupSafe==3.0.3
mashumaro==3.14
more-itertools==10.8.0
msgpack==1.1.2
narwhals==2.11.0
networkx==3.5
numpy==2.3.4
oauthlib==3.3.1
orderly-set==5.5.0
packaging==25.0
pandas==2.3.3
parsedatetime==2.6
pathspec==0.12.1
pillow==12.0.0
platformdirs==4.5.0
protobuf==6.33.1
psycopg2-binary==2.9.11
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.4
pydantic_core==2.41.5
pydeck==0.9.1
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
requests-file==3.0.1
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rpds-py==0.28.0
rsa==4.9.1
six==1.17.0
smmap==5.0.2
snowplow-tracker==1.1.0
SQLAlchemy==2.0.44
sqlparse==0.5.3
streamlit==1.51.0
tenacity==9.1.2
text-unidecode==1.3
toml==0.10.2
tornado==6.5.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
xmltodict==0.15.1
zeep==4.3.2
zipp==3.23.0
